# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "FlexPrice API Service",
        "title": "FlexPrice API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
...
        "/meters": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all meters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "List meters",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.MeterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Create a new meter with the specified configuration",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Create meter",
                "parameters": [
                    {
                        "description": "Meter configuration",
                        "name": "meter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateMeterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.MeterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/meters/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get a specific meter by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Get meter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MeterResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update an existing meter",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Update meter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Meter configuration",
                        "name": "meter",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateMeterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.MeterResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete an existing meter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Delete meter",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message:Meter deleted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/meters/{id}/disable": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Disable an existing meter",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meters"
                ],
                "summary": "Disable meter [TODO: Deprecate]",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Meter ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message:Meter disabled successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateMeterRequest": {
            "type": "object",
            "required": [
                "aggregation",
                "event_name",
                "name"
            ],
            "properties": {
                "aggregation": {
                    "$ref": "#/definitions/meter.Aggregation"
                },
                "event_name": {
                    "type": "string",
                    "example": "api_request"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meter.Filter"
                    }
                },
                "name": {
                    "type": "string",
                    "example": "API Usage Meter"
                },
                "reset_usage": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.ResetUsage"
                        }
                    ],
                    "example": "BILLING_PERIOD"
                }
            }
        },
...
        "dto.MeterResponse": {
            "type": "object",
            "properties": {
                "aggregation": {
                    "$ref": "#/definitions/meter.Aggregation"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-03-20T15:04:05Z"
                },
                "event_name": {
                    "type": "string",
                    "example": "api_request"
                },
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meter.Filter"
                    }
                },
                "id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "name": {
                    "type": "string",
                    "example": "API Usage Meter"
                },
                "reset_usage": {
                    "$ref": "#/definitions/types.ResetUsage"
                },
                "status": {
                    "type": "string",
                    "example": "published"
                },
                "tenant_id": {
                    "type": "string",
                    "example": "tenant123"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-03-20T15:04:05Z"
                }
            }
        },
...
        "dto.UpdateMeterRequest": {
            "type": "object",
            "properties": {
                "filters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/meter.Filter"
                    }
                }
            }
        },
...
        "meter.Aggregation": {
            "type": "object",
            "properties": {
                "field": {
                    "description": "Field is the key in $event.properties on which the aggregation is to be applied\nFor ex if the aggregation type is sum for API usage, the field could be \"duration_ms\"",
                    "type": "string"
                },
                "type": {
                    "description": "Type is the type of aggregation to be applied on the events\nFor ex sum, count, avg, max, min etc",
                    "allOf": [
                        {
                            "$ref": "#/definitions/types.AggregationType"
                        }
                    ]
                }
            }
        },
        "meter.Filter": {
            "type": "object",
            "properties": {
                "key": {
                    "description": "Key is the key for the filter from $event.properties\nCurrently we support only first level keys in the properties and not nested keys",
                    "type": "string"
                },
                "values": {
                    "description": "Values are the possible values for the filter to be considered for the meter\nFor ex \"model_name\" could have values \"o1-mini\", \"gpt-4o\" etc",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "types.AggregationType": {
            "type": "string",
            "enum": [
                "COUNT",
                "SUM",
                "AVG"
            ],
            "x-enum-varnames": [
                "AggregationCount",
                "AggregationSum",
                "AggregationAvg"
            ]
        },
...
        "types.ResetUsage": {
            "type": "string",
            "enum": [
                "BILLING_PERIOD",
                "NEVER"
            ],
            "x-enum-varnames": [
                "ResetUsageBillingPeriod",
                "ResetUsageNever"
            ]
        },
...
        "v1.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string",
                    "example": "Invalid request payload"
                },
                "error": {
                    "type": "string",
                    "example": "Invalid request payload"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Enter your API key in the format *x-api-key \u0026lt;api-key\u0026gt;**",
            "type": "apiKey",
            "name": "x-api-key",
            "in": "header"
        }
    }
}